decouple main routing implementation from routing package backend
