+++ src/nwm_routing/pyproject.toml
@@ -1,0 +1,8 @@
+[build-system]
+requires = [
+    "setuptools>=42",
+    "wheel",
+    "pip",
+    "Cython"
+]
+build-backend = "setuptools.build_meta"
