+++ src/nwm_routing/setup.cfg
@@ -1,0 +1,20 @@
+[metadata]
+name=nwm_routing
+author=James Halgren
+author_email=james.halgren@noaa.gov
+
+[options]
+package_dir=
+  =src
+packages=find:
+python_requires=>=3.6
+install_requires=
+  numpy
+  Cython
+  pandas
+  geopandas
+  xarray
+  netcdf4
+  pyyaml
+  toolz
+  joblib
