+++ src/nwm_routing/setup.py
@@ -1,0 +1,4 @@
+#"install" the nwm_routing package/modules
+from setuptools  import setup
+
+setup()
