+++ src/nwm_routing/src/nwm_routing/input.py
@@ -1,0 +1,150 @@
+import pathlib
+
+import troute.nhd_network_utilities_v02 as nnu
+import troute.nhd_io as nhd_io
+
+#FIXME
+ENV_IS_CL = False
+if ENV_IS_CL:
+    root = pathlib.Path("/", "content", "t-route")
+elif not ENV_IS_CL:
+    root = pathlib.Path("../../").resolve()
+
+def _input_handler(args):
+
+    #args = _handle_args()
+
+    custom_input_file = args.custom_input_file
+    supernetwork_parameters = None
+    waterbody_parameters = {}
+    forcing_parameters = {}
+    restart_parameters = {}
+    output_parameters = {}
+    run_parameters = {}
+    parity_parameters = {}
+    data_assimilation_parameters = {}
+    diffusive_parameters = {}
+
+    if custom_input_file:
+        (
+            supernetwork_parameters,
+            waterbody_parameters,
+            forcing_parameters,
+            restart_parameters,
+            output_parameters,
+            run_parameters,
+            parity_parameters,
+            data_assimilation_parameters,
+            diffusive_parameters,
+        ) = nhd_io.read_custom_input(custom_input_file)
+
+    else:
+        run_parameters["assume_short_ts"] = args.assume_short_ts
+        run_parameters["return_courant"] = args.return_courant
+        run_parameters["parallel_compute_method"] = args.parallel_compute_method
+        run_parameters["subnetwork_target_size"] = args.subnetwork_target_size
+        run_parameters["cpu_pool"] = args.cpu_pool
+        run_parameters["showtiming"] = args.showtiming
+        run_parameters["compute_method"] = args.compute_method
+
+        run_parameters["debuglevel"] = debuglevel = -1 * args.debuglevel
+        run_parameters["verbose"] = verbose = args.verbose
+
+        output_parameters["csv_output"] = {}
+        output_parameters["csv_output"]["csv_output_folder"] = args.csv_output_folder
+
+        test_folder = pathlib.Path(root, "test")
+        geo_input_folder = test_folder.joinpath("input", "geo")
+
+        test_case = args.test_case
+
+        if test_case:
+
+            # call test case assemble function
+            (
+                supernetwork_parameters,
+                run_parameters,
+                output_parameters,
+                restart_parameters,
+                forcing_parameters,
+                parity_parameters,
+            ) = build_tests.build_test_parameters(
+                test_case,
+                supernetwork_parameters,
+                run_parameters,
+                output_parameters,
+                restart_parameters,
+                forcing_parameters,
+                parity_parameters,
+            )
+
+        else:
+            run_parameters["dt"] = args.dt
+            run_parameters["nts"] = args.nts
+            run_parameters["qts_subdivisions"] = args.qts_subdivisions
+            run_parameters["compute_method"] = args.compute_method
+
+            waterbody_parameters[
+                "break_network_at_waterbodies"
+            ] = args.break_network_at_waterbodies
+
+            data_assimilation_parameters[
+                "data_assimilation_parameters_folder"
+            ] = args.data_assimilation_parameters_folder
+            data_assimilation_parameters[
+                "data_assimilation_filter"
+            ] = args.data_assimilation_filter
+            data_assimilation_parameters[
+                "data_assimilation_csv"
+            ] = args.data_assimilation_csv
+
+            restart_parameters[
+                "wrf_hydro_channel_restart_file"
+            ] = args.wrf_hydro_channel_restart_file
+            restart_parameters[
+                "wrf_hydro_channel_ID_crosswalk_file"
+            ] = args.wrf_hydro_channel_ID_crosswalk_file
+            restart_parameters[
+                "wrf_hydro_channel_ID_crosswalk_file_field_name"
+            ] = args.wrf_hydro_channel_ID_crosswalk_file_field_name
+            restart_parameters[
+                "wrf_hydro_channel_restart_upstream_flow_field_name"
+            ] = args.wrf_hydro_channel_restart_upstream_flow_field_name
+            restart_parameters[
+                "wrf_hydro_channel_restart_downstream_flow_field_name"
+            ] = args.wrf_hydro_channel_restart_downstream_flow_field_name
+            restart_parameters[
+                "wrf_hydro_channel_restart_depth_flow_field_name"
+            ] = args.wrf_hydro_channel_restart_depth_flow_field_name
+
+            forcing_parameters["qlat_const"] = float(args.qlat_const)
+            forcing_parameters["qlat_input_folder"] = args.qlat_input_folder
+            forcing_parameters["qlat_input_file"] = args.qlat_input_file
+            forcing_parameters[
+                "qlat_file_pattern_filter"
+            ] = args.qlat_file_pattern_filter
+            forcing_parameters["qlat_file_index_col"] = args.qlat_file_index_col
+            forcing_parameters["qlat_file_value_col"] = args.qlat_file_value_col
+
+            supernetwork = args.supernetwork
+
+        # STEP 0.5: Obtain Supernetwork Parameters for test cases
+        if not supernetwork_parameters:
+            supernetwork_parameters = nnu.set_supernetwork_parameters(
+                supernetwork=supernetwork,
+                geo_input_folder=geo_input_folder,
+                verbose=False,
+                debuglevel=debuglevel,
+            )
+
+    return (
+        supernetwork_parameters,
+        waterbody_parameters,
+        forcing_parameters,
+        restart_parameters,
+        output_parameters,
+        run_parameters,
+        parity_parameters,
+        data_assimilation_parameters,
+        diffusive_parameters,
+    )
