+++ src/python_routing_v02/troute/routing/utils.py
@@ -1,0 +1,11 @@
+import pandas as pd
+import numpy as np
+
+def writetoFile(file, writeString):
+    file.write(writeString)
+    file.write("\n")
+
+def constant_qlats(index_dataset, nsteps, qlat):
+    q = np.full((len(index_dataset.index), nsteps), qlat, dtype="float32")
+    ql = pd.DataFrame(q, index=index_dataset.index, columns=range(nsteps))
+    return ql
